requirements.txt
flask
numpy
gunicorn
google_trans_new
opencv-python
gensim
io
Pillow
tensorflow
